test(file): Move datetime tests to their formats
This will make it easier to remember to add them when new formats are
added
