+++ tests/testsuite/file_json5.rs
@@ -2,9 +2,11 @@
 
 use std::path::PathBuf;
 
-use config::{Config, File, FileFormat, Map, Value};
+use chrono::{DateTime, TimeZone, Utc};
 use float_cmp::ApproxEqUlps;
 use serde_derive::Deserialize;
+
+use config::{Config, File, FileFormat, Map, Value};
 
 #[test]
 fn test_file() {
@@ -200,4 +202,24 @@ fn test_override_lowercase_value_for_enums() {
         param,
         EnumSettings::Bar("I have been overridden_with_lower_case".to_owned())
     );
+}
+
+#[test]
+fn json() {
+    let s = Config::builder()
+        .add_source(File::from_str(
+            r#"
+            {
+                "json_datetime": "2017-05-10T02:14:53Z"
+            }
+            "#,
+            FileFormat::Json5,
+        ))
+        .build()
+        .unwrap();
+
+    let date: String = s.get("json_datetime").unwrap();
+    assert_eq!(&date, "2017-05-10T02:14:53Z");
+    let date: DateTime<Utc> = s.get("json_datetime").unwrap();
+    assert_eq!(date, Utc.with_ymd_and_hms(2017, 5, 10, 2, 14, 53).unwrap());
 }
